Make the handler selectable
