+++ src/handlers/index.ts
@@ -1,0 +1,23 @@
+import { Chrome111 } from './Chrome111';
+import { Chrome74 } from './Chrome74';
+import { Chrome70 } from './Chrome70';
+import { Chrome67 } from './Chrome67';
+import { Chrome55 } from './Chrome55';
+import { Firefox60 } from './Firefox60';
+import { Safari12 } from './Safari12';
+import { Safari11 } from './Safari11';
+import { Edge11 } from './Edge11';
+import { ReactNativeUnifiedPlan } from './ReactNativeUnifiedPlan';
+import { ReactNative } from './ReactNative';
+
+export { Chrome111
+	, Chrome74
+	, Chrome70
+	, Chrome67
+	, Chrome55
+	, Firefox60
+	, Safari12
+	, Safari11
+	, Edge11
+	, ReactNativeUnifiedPlan
+	, ReactNative };
