+++ src/index.ts
@@ -1,5 +1,7 @@
 import debug from 'debug';
-import { Device, detectDevice } from './Device';
+import { Device, detectDevice, ManualDevice } from './Device';
+
+
 import * as types from './types';
 
 /**
@@ -15,7 +17,9 @@ export const version = '__MEDIASOUP_CLIENT_VERSION__';
 /**
  * Expose Device class and detectDevice() helper.
  */
-export { Device, detectDevice };
+export { Device, detectDevice, ManualDevice };
+
+export { builtinHandlerFactory, browserHandlerFactory, reactNativeHandlerFactory } from './BuiltinHandlerFactory';
 
 /**
  * Expose parseScalabilityMode() function.
